calculator/requirements.txt
Appium-Python-Client==0.24
selenium==3.5.0
webdriver-manager
pytest
pytest-html